% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{classicCount-class}
\alias{GOFisherTest,classicCount-method}
\alias{classicCount-class}
\alias{contTable}
\alias{contTable,classicCount-method}
\alias{initialize,classicCount-method}
\alias{numSigAll}
\alias{numSigAll,classicCount-method}
\alias{numSigMembers}
\alias{numSigMembers,classicCount-method}
\alias{sigAllMembers}
\alias{sigAllMembers,classicCount-method}
\alias{sigMembers}
\alias{sigMembers,classicCount-method}
\alias{sigMembers<-}
\alias{sigMembers<-,classicCount-method}
\title{Class "classicCount"}
\description{
This class that extends the virtual class "groupStats"
 by adding a slot representing the significant members.
}
\details{
This class is used for test statistic based on counts, like Fisher's
 exact test
}
\section{Objects from the Class}{

 Objects can be created by calls of the form
 \code{new("classicCount",
         testStatistic = "function",
         name = "character",
         allMembers = "character",
         groupMembers = "character",
         sigMembers = "character")}.
}

\section{Slots}{

\describe{
 \item{significant}{Object of class \code{"integer"}}
 \item{name}{Object of class \code{"character"}}
 \item{allMembers}{Object of class \code{"character"}}
 \item{members}{Object of class \code{"character"}}
 \item{testStatistic}{Object of class \code{"function"}}
  }
}

\section{Extends}{

 Class \code{"groupStats"}, directly.
}

\section{Methods}{

 \describe{
   \item{contTable}{\code{signature(object = "classicCount")}: ... }
   \item{initialize}{\code{signature(.Object = "classicCount")}: ... }
   \item{numSigAll}{\code{signature(object = "classicCount")}: ... }
   \item{numSigMembers}{\code{signature(object = "classicCount")}: ... }
   \item{sigAllMembers}{\code{signature(object = "classicCount")}: ... }
   \item{sigMembers<-}{\code{signature(object = "classicCount")}: ... }
   \item{sigMembers}{\code{signature(object = "classicCount")}: ... }
 }
}
\examples{
data(ONTdata)
gene.universe <- genes(GOdata)
go.genes <- genesInTerm(GOdata, 'DOID:0050709')[[1]]
sig.genes <- sigGenes(GOdata)
my.group <- new("classicCount", testStatistic = GOFisherTest, name = "fisher",allMembers = gene.universe, groupMembers = go.genes,sigMembers = sig.genes)
contTable(my.group)
runTest(my.group)
}
\author{
Adrian Alexa
}
\seealso{
\code{\link{classicScore-class}},
   \code{\link{groupStats-class}},
   \code{\link{getSigGroups-methods}}
}
\keyword{classes}

