% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{elimCount-class}
\alias{GOFisherTest,elimCount-method}
\alias{contTable,elimCount-method}
\alias{cutOff}
\alias{cutOff,elimCount-method}
\alias{cutOff<-}
\alias{cutOff<-,elimCount-method}
\alias{depth}
\alias{depth,leaCount-method}
\alias{depth<-}
\alias{depth<-,leaCount-method}
\alias{elim}
\alias{elim,elimCount-method}
\alias{elim,weight01Count-method}
\alias{elim<-}
\alias{elim<-,elimCount-method}
\alias{elim<-,weight01Count-method}
\alias{elimCount-class}
\alias{initialize,elimCount-method}
\alias{initialize,leaCount-method}
\alias{initialize,weight01Count-method}
\alias{leaCount-class}
\alias{numAllMembers,elimCount-method}
\alias{numAllMembers,weight01Count-method}
\alias{numMembers,elimCount-method}
\alias{numMembers,weight01Count-method}
\alias{numSigAll,elimCount-method}
\alias{numSigAll,weight01Count-method}
\alias{numSigMembers,elimCount-method}
\alias{numSigMembers,weight01Count-method}
\alias{sigAllMembers,elimCount-method}
\alias{sigAllMembers,weight01Count-method}
\alias{sigMembers,elimCount-method}
\alias{sigMembers,weight01Count-method}
\alias{sigMembers<-,elimCount-method}
\alias{weight01Count-class}
\title{Classes "elimCount" and "weight01Count"}
\description{
Classes that extend the "classicCount" class by adding a slot representing the members that need to be removed.
}
\section{Objects from the Class}{

  Objects can be created by calls of the form \code{new("elimCount", testStatistic, name, allMembers, groupMembers, sigMembers, elim, cutOff, ...)}.
}

\section{Slots}{

  \describe{
    \item{\code{elim}:}{Object of class \code{"integer"} ~~ }
    \item{\code{cutOff}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{significant}:}{Object of class \code{"integer"} ~~ }
    \item{\code{name}:}{Object of class \code{"character"} ~~ }
    \item{\code{allMembers}:}{Object of class \code{"character"} ~~ }
    \item{\code{members}:}{Object of class \code{"character"} ~~ }
    \item{\code{testStatistic}:}{Object of class \code{"function"} ~~ }
    \item{\code{testStatPar}:}{Object of class \code{"list"} ~~ }
  }
}

\section{Extends}{

  Class \code{"\linkS4class{classicCount}"}, directly.
Class \code{"\linkS4class{groupStats}"}, by class "classicCount", distance 2.
}

\section{Methods}{

  No methods defined with class "elimCount" in the signature.
}
\examples{
data(ONTdata)
gene.universe <- genes(GOdata)
go.genes <- genesInTerm(GOdata, 'DOID:0050709')[[1]]
elim.genes <- sample(go.genes, length(go.genes) / 4)
sig.genes <- sigGenes(GOdata)
elim.group <- new("elimCount", testStatistic = GOFisherTest, name = "fisher",allMembers = gene.universe, groupMembers = go.genes,sigMembers = sig.genes, elim = elim.genes)
contTable(elim.group)
runTest(elim.group)
}
\author{
Adrian Alexa
}
\seealso{
\code{\link{classicScore-class}},
  \code{\link{groupStats-class}},
  \code{\link{getSigGroups-methods}}
}
\keyword{classes}

