% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{elimExpr-class}
\alias{allMembers,weight01Expr-method}
\alias{cutOff,elimExpr-method}
\alias{cutOff<-,elimExpr-method}
\alias{depth,leaExpr-method}
\alias{depth<-,leaExpr-method}
\alias{elim,weight01Expr-method}
\alias{elim<-,weight01Expr-method}
\alias{elimExpr-class}
\alias{initialize,elimExpr-method}
\alias{initialize,leaExpr-method}
\alias{initialize,weight01Expr-method}
\alias{leaExpr-class}
\alias{members,weight01Expr,missing-method}
\alias{numAllMembers,weight01Expr-method}
\alias{numMembers,weight01Expr-method}
\alias{weight01Expr-class}
\title{Class "elimExpr"}
\description{
Classes that extend the "classicExpr" class by adding a slot representing the members that need to be removed.
}
\section{Objects from the Class}{

Objects can be created by calls of the form \code{new("elimExpr", testStatistic, name, groupMembers, exprDat, pType, elim, cutOff, ...)}.
   ~~ describe objects here ~~
}

\section{Slots}{

 \describe{
    \item{\code{cutOff}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{elim}:}{Object of class \code{"integer"} ~~ }
    \item{\code{eData}:}{Object of class \code{"environment"} ~~ }
    \item{\code{pType}:}{Object of class \code{"factor"} ~~ }
    \item{\code{name}:}{Object of class \code{"character"} ~~ }
    \item{\code{allMembers}:}{Object of class \code{"character"} ~~ }
    \item{\code{members}:}{Object of class \code{"character"} ~~ }
    \item{\code{testStatistic}:}{Object of class \code{"function"} ~~ }
    \item{\code{testStatPar}:}{Object of class \code{"list"} ~~ }
  }
}

\section{Extends}{

Class \code{"\linkS4class{weight01Expr}"}, directly.
Class \code{"\linkS4class{classicExpr}"}, by class "weight01Expr", distance 2.
Class \code{"\linkS4class{groupStats}"}, by class "weight01Expr", distance 3.
}

\section{Methods}{

  \describe{
    \item{cutOff<-}{\code{signature(object = "elimExpr")}: ... }
    \item{cutOff}{\code{signature(object = "elimExpr")}: ... }
    \item{getSigGroups}{\code{signature(object = "topONTdata", test.stat = "elimExpr")}: ... }
    \item{initialize}{\code{signature(.Object = "elimExpr")}: ... }
 }
}
\examples{
showClass("elimExpr")
}
\author{
Adrian Alexa
}
\seealso{
\code{\link{classicScore-class}},
  \code{\link{groupStats-class}},
  \code{\link{getSigGroups-methods}}
}
\keyword{classes}

