% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/topGOtests.R
\name{Gene set tests statistics}
\alias{GOFisherTest}
\alias{GOKSTest}
\alias{GOKSTiesTest}
\alias{GOSumTest}
\alias{GOglobalTest}
\alias{GOtTest}
\alias{Gene set tests statistics}
\alias{permSumStats}
\alias{permSumStats.all}
\title{Gene set tests statistics}
\usage{
GOFisherTest(object)
GOKSTest(object)
GOtTest(object)
GOglobalTest(object)
GOSumTest(object)
GOKSTiesTest(object)
}
\arguments{
\item{object}{An object of class \code{groupStats} or decedent class.}
}
\value{
All these methods return the p-value computed by the respective test statistic.
}
\description{
Methods which implement and run a group test statistic for
  a class inheriting from \code{groupStats} class. See Details section
  for a description of each method.
}
\details{
GOFisherTest: implements Fischer's exact test (based on contingency
  table) for \code{groupStats} objects dealing with "counts".

  GOKSTest: implements the Kolmogorov-Smirnov test for \code{groupStats}
  objects dealing with gene "scores". This test uses the \code{ks.test}
  function and does not implement the running-sum-statistic test based
  on permutations.

  GOtTest: implements the t-test for \code{groupStats} objects dealing
  with gene "scores". It should be used when the gene scores are
  t-statistics or any other score following a normal distribution.

  GOglobalTest: implement Goeman's globaltest.
}
\examples{
data(ONTdata)
gene.universe <- genes(GOdata)
go.genes <- genesInTerm(GOdata, 'DOID:0050709')[[1]]
sig.genes <- sigGenes(GOdata)
my.group <- new("classicCount", testStatistic = GOFisherTest, name = "fisher",allMembers = gene.universe, groupMembers = go.genes,sigMembers = sig.genes)
contTable(my.group)
runTest(my.group)
}
\author{
Adrian Alexa
}
\seealso{
\code{\link{groupStats-class}},
  \code{\link{getSigGroups-methods}}
}
\keyword{misc}

