% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/topGOgraph.R
\name{inducedGraph}
\alias{inducedGraph}
\alias{nodesInInducedGraph}
\title{The subgraph induced by a set of nodes.}
\usage{
inducedGraph(dag, startNodes)
nodesInInducedGraph(dag, startNodes)
}
\arguments{
\item{dag}{An object of class \code{graphNEL} containing a directed graph.}

\item{startNodes}{A character vector giving the starting nodes.}
}
\value{
An object of class \code{\link[graph]{graphNEL-class}} is returned.
}
\description{
Given a set of nodes (GO terms) this function is returning
  the subgraph containing these nodes and their ancestors.
}
\examples{
data(ONTdata)

## the GO graph
g <- graph(GOdata)
g

## select 10 random nodes
sn <- sample(nodes(g), 10)


## the subgraph induced by these nodes
sg <- inducedGraph(g, sn)
sg
}
\author{
Adrian Alexa#'
}
\seealso{
\code{\link{topONTdata-class}},
  \code{\link{reverseArch}},
}
\keyword{graphs}

