% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/topGOfunctions.R
\name{init}
\alias{init}
\alias{initONT}
\alias{initWHAT}
\title{init methods}
\usage{
initWHAT()
initONT(ontology)
}
\arguments{
\item{ontology}{the ontology to be loaded. This can be check by running initWHAT()}
}
\description{
functions to load the ontology objects
}
\details{
initONT loads the ontology object from topOnto.<ontolog>.db packages.
}
\examples{
initWHAT()
\dontrun{
 require(topOnto.HDO.db)
 initONT('HDO')
}
}
\author{
Xin He
}

