% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{topONTdata-class}
\alias{allGenes}
\alias{allGenes,topONTdata-method}
\alias{attrInTerm}
\alias{attrInTerm,topONTdata,character,character-method}
\alias{attrInTerm,topONTdata,character,missing-method}
\alias{countGenesInTerm}
\alias{countGenesInTerm,topONTdata,character-method}
\alias{countGenesInTerm,topONTdata,missing-method}
\alias{description}
\alias{description,topONTdata-method}
\alias{description<-}
\alias{description<-,topONTdata,ANY-method}
\alias{expressionMatrix}
\alias{expressionMatrix,topONTdata-method}
\alias{feasible}
\alias{feasible,topONTdata-method}
\alias{feasible<-}
\alias{feasible<-,topONTdata-method}
\alias{geneScore}
\alias{geneScore,topONTdata,character-method}
\alias{geneScore,topONTdata,missing-method}
\alias{geneScore,topONTdata-method}
\alias{geneSelectionFun}
\alias{geneSelectionFun,topONTdata-method}
\alias{geneSelectionFun<-}
\alias{geneSelectionFun<-,topONTdata-method}
\alias{genes}
\alias{genes,topONTdata-method}
\alias{genesInTerm}
\alias{genesInTerm,topONTdata,character-method}
\alias{genesInTerm,topONTdata,missing-method}
\alias{graph}
\alias{graph,topONTdata-method}
\alias{graph<-}
\alias{graph<-,topONTdata-method}
\alias{initialize,topONTdata-method}
\alias{numGenes}
\alias{numGenes,topONTdata-method}
\alias{numSigGenes}
\alias{numSigGenes,topONTdata-method}
\alias{ontology}
\alias{ontology,topONTdata-method}
\alias{ontology<-}
\alias{ontology<-,topONTdata-method}
\alias{phenotype}
\alias{phenotype,topONTdata-method}
\alias{print,topONTdata-method}
\alias{scoresInTerm}
\alias{scoresInTerm,topONTdata,character-method}
\alias{scoresInTerm,topONTdata,missing-method}
\alias{show,topONTdata-method}
\alias{sigGenes}
\alias{sigGenes,topONTdata-method}
\alias{termStat}
\alias{termStat,topONTdata,character-method}
\alias{termStat,topONTdata,missing-method}
\alias{topONTdata-class}
\alias{updateGenes}
\alias{updateGenes,topONTdata,factor,missing-method}
\alias{updateGenes,topONTdata,numeric,function-method}
\alias{updateTerm<-}
\alias{updateTerm<-,topONTdata,character-method}
\alias{usedGO}
\alias{usedGO,topONTdata-method}
\title{Class "topONTdata"}
\description{
TODO: The node attributes are environments containing the genes/probes
  annotated to the respective node

  If genes is a numeric vector than this should represent the gene's
  score. If it is factor it should discriminate the genes in
  interesting genes and the rest

  TODO: it will be a good idea to replace the allGenes and allScore with an
  ExpressionSet class. In this way we can use tests like global test, globalAncova....
  -- ALL variables starting with . are just for internal class usage (private)
}
\section{Objects from the Class}{

  Objects can be created by calls of the form \code{new("topONTdata", ontology, allGenes, geneSelectionFun, description, annotationFun, ...)}.
  ~~ describe objects here ~~
}

\section{Slots}{

  \describe{
    \item{\code{description}:}{Object of class \code{"character"} ~~ }
    \item{\code{ontology}:}{Object of class \code{"character"} ~~ }
    \item{\code{allGenes}:}{Object of class \code{"character"} ~~ }
    \item{\code{allScores}:}{Object of class \code{"ANY"} ~~ }
    \item{\code{geneSelectionFun}:}{Object of class \code{"function"} ~~ }
    \item{\code{feasible}:}{Object of class \code{"logical"} ~~ }
    \item{\code{nodeSize}:}{Object of class \code{"integer"} ~~ }
    \item{\code{graph}:}{Object of class \code{"graphNEL"} ~~ }
    \item{\code{expressionMatrix}:}{Object of class \code{"matrix"} ~~ }
    \item{\code{phenotype}:}{Object of class \code{"factor"} ~~ }

  }
}

\section{Methods}{

  \describe{
    \item{allGenes}{\code{signature(object = "topONTdata")}: ... }
    \item{attrInTerm}{\code{signature(object = "topONTdata", attr = "character", whichGO = "character")}: ... }
    \item{attrInTerm}{\code{signature(object = "topONTdata", attr = "character", whichGO = "missing")}: ... }
    \item{countGenesInTerm}{\code{signature(object = "topONTdata", whichGO = "character")}: ... }
    \item{countGenesInTerm}{\code{signature(object = "topONTdata", whichGO = "missing")}: ... }
    \item{description<-}{\code{signature(object = "topONTdata")}: ... }
    \item{description}{\code{signature(object = "topONTdata")}: ... }
    \item{feasible<-}{\code{signature(object = "topONTdata")}: ... }
    \item{feasible}{\code{signature(object = "topONTdata")}: ... }
    \item{geneScore}{\code{signature(object = "topONTdata")}: ... }
    \item{geneSelectionFun<-}{\code{signature(object = "topONTdata")}: ... }
    \item{geneSelectionFun}{\code{signature(object = "topONTdata")}: ... }

    \item{genes}{\code{signature(object = "topONTdata")}: A method for
      obtaining the list of genes, as a characther vector, which will be
      used in the further analysis.}
    \item{numGenes}{\code{signature(object = "topONTdata")}: A method for
      obtaining the number of genes, which will be used in the further
      analysis. It has the same effect as: \code{lenght(genes(object))}.}

    \item{sigGenes}{\code{signature(object = "topONTdata")}:  A method for
      obtaining the list of significant genes, as a charachter vector.}

    \item{genesInTerm}{\code{signature(object = "topONTdata", whichGO = "character")}: ... }
    \item{genesInTerm}{\code{signature(object = "topONTdata", whichGO = "missing")}: ... }
    \item{getSigGroups}{\code{signature(object = "topONTdata", test.stat = "classicCount")}: ... }
    \item{getSigGroups}{\code{signature(object = "topONTdata", test.stat = "classicScore")}: ... }
    \item{graph<-}{\code{signature(object = "topONTdata")}: ... }
    \item{graph}{\code{signature(object = "topONTdata")}: ... }
    \item{initialize}{\code{signature(.Object = "topONTdata")}: ... }

    \item{ontology<-}{\code{signature(object = "topONTdata")}: ... }
    \item{ontology}{\code{signature(object = "topONTdata")}: ... }

    \item{termStat}{\code{signature(object = "topONTdata", whichGO = "character")}: ... }
    \item{termStat}{\code{signature(object = "topONTdata", whichGO = "missing")}: ... }
    \item{updateGenes}{\code{signature(object = "topONTdata", geneList = "numeric", geneSelFun = "function")}: ... }
    \item{updateGenes}{\code{signature(object = "topONTdata", geneList = "factor", geneSelFun = "missing")}: ... }
    \item{updateTerm<-}{\code{signature(object = "topONTdata", attr = "character")}: ... }
    \item{usedGO}{\code{signature(object = "topONTdata")}: ... }
  }
}
\examples{
## load the dataset
data(ONTdata)
#library(package = affyLib, character.only = TRUE)

## the distribution of the adjusted p-values
#hist(geneList, 100)

## how many differentially expressed genes are:
#sum(topDiffGenes(geneList))

## build the topONTdata class
\dontrun{
require(topOnto.HDO.db)
GOdata <- new("topONTdata", ontology = "HDO", description="HDO example",allGenes = geneList,annot = annFUN.gene2GO, gene2GO = geneID2TERM)
## display the GOdata object
GOdata
}

##########################################################
## Examples on how to use the methods
##########################################################

## description of the experiment
description(GOdata)

## obtain the genes that will be used in the analysis
a <- genes(GOdata)
str(a)
numGenes(GOdata)

## obtain the score (p-value) of the genes
selGenes <- names(geneList)[sample(1:length(geneList), 10)]
gs <- geneScore(GOdata, whichGenes = selGenes)
print(gs)

## if we want an unnamed vector containing all the feasible genes
gs <- geneScore(GOdata, use.names = FALSE)
str(gs)

## the list of significant genes
sg <- sigGenes(GOdata)
str(sg)
numSigGenes(GOdata)

## to update the gene list
.geneList <- geneScore(GOdata, use.names = TRUE)
GOdata ## more available genes
#GOdata <- updateGenes(GOdata, .geneList, topDiffGenes)
GOdata ## the available genes are now the feasible genes

## the available GO terms (all the nodes in the graph)
go <- usedGO(GOdata)
length(go)

## to list the genes annotated to a set of specified GO terms
sel.terms <- sample(go, 10)
ann.genes <- genesInTerm(GOdata, sel.terms)
str(ann.genes)

## the score for these genes
ann.score <- scoresInTerm(GOdata, sel.terms)
str(ann.score)

## to see the number of annotated genes
num.ann.genes <- countGenesInTerm(GOdata)
str(num.ann.genes)

## to summarise the statistics
termStat(GOdata, sel.terms)
}
\author{
Adrian Alexa
}
\seealso{
\code{\link{buildLevels}},
  \code{\link{annFUN}}
}
\keyword{graphs,classes}

